いいね機能に色変更機能を追加,ユーザー登録機能にバリデーションを追加,全体的な画面遷移、画面デザインの整理

+++ src/main/java/com/SoupOfYingyme/QuestionDataService.java
@@ -41,7 +41,7 @@ public class QuestionDataService {
 	
 	public void addGood(Long id) {
 		Optional<QuestionData> questiondata = repository.findById(id);
-		int goodpoint = questiondata.get().getGood();
+		long goodpoint = questiondata.get().getGood();
 		goodpoint++;
 		questiondata.get().setGood(goodpoint);
 		repository.saveAndFlush(questiondata.get());
